team-info: move from settings to account
