+++ packages/pilot/src/containers/Settings/User/TeamInfoTab/TeamManagement/RemoveUserModal.js
